requirements.txt
Flask==2.3.3
flask-login==0.6.2
twilio==8.10.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
bcrypt==4.0.1
APScheduler==3.10.4
python-dateutil==2.8.2 